test: scrap aliases 테스트 추가
